+++ four/src/lambda/event_invoke_config.rs
@@ -8,9 +8,3 @@ pub struct MaximumEventAgeInSeconds(usize);
 
 #[nutype(validate(less_or_equal = 2), derive(Debug, Clone, Serialize))]
 pub struct MaximumRetryAttempts(usize);
-
-#[nutype(
-    validate(regex = r#"^(|[a-zA-Z0-9$_-]{1,129})$"#),
-    derive(Debug, Clone, Serialize)
-)]
-pub struct Qualifier(String);
