+++ four/src/lambda/id.rs
@@ -1,2 +1,5 @@
 #[derive(Debug, Clone)]
 pub struct LayerVersionPermissionId;
+
+#[derive(Debug, Clone)]
+pub struct UrlId;
