+++ four/src/lambda/mod.rs
@@ -5,12 +5,15 @@ mod function;
 mod id;
 mod layer_version;
 mod layer_version_permission;
+mod permission;
 mod property;
 pub mod resource;
+mod url;
+mod version;
 
 pub use alias::{
     AliasDescription, AliasDescriptionError, AliasName, AliasNameError, AliasRoutingConfiguration,
-    FunctionVersion, FunctionVersionError, ProvisionedConcurrencyConfiguration, VersionWeight,
+    FunctionVersion, FunctionVersionError, VersionWeight,
 };
 pub use arn::FunctionArn;
 pub use event_invoke_config::{
@@ -20,7 +23,7 @@ pub use event_invoke_config::{
 pub use function::{
     FunctionName, FunctionNameError, MemorySize, MemorySizeError, Timeout, TimeoutError,
 };
-pub use id::LayerVersionPermissionId;
+pub use id::{LayerVersionPermissionId, UrlId};
 pub use layer_version::{
     CompatibleArchitectures, CompatibleArchitecturesError, FunctionContent, LayerName,
     LayerNameError, LayerVersionDescription, LayerVersionDescriptionError, LicenseInfo,
@@ -29,6 +32,10 @@ pub use layer_version::{
 pub use layer_version_permission::{
     GetLayerVersionAction, OrganizationId, OrganizationIdError, Principal, PrincipalError,
 };
+pub use permission::{
+    EventSourceToken, EventSourceTokenError, PermissionPrincipal, PrincipalOrgID,
+    PrincipalOrgIDError,
+};
 pub use property::{
     architecture::{Architecture, Architectures},
     code::Code,
@@ -36,3 +43,37 @@ pub use property::{
     handler::{Handler, HandlerError},
     runtime::Runtime,
 };
+use serde::Serialize;
+pub use url::{
+    AuthType, Cors, Header, HeaderError, InvokeMode, MaxAge, MaxAgeError, Method, Origin,
+};
+pub use version::{
+    RuntimePolicy, RuntimeVersionArn, RuntimeVersionArnError, VersionDescription,
+    VersionDescriptionError,
+};
+
+#[derive(Debug, Clone, Serialize)]
+#[serde(rename_all = "PascalCase")]
+pub struct ProvisionedConcurrencyConfiguration {
+    provisioned_concurrency_executions: usize,
+}
+
+#[derive(Debug, Clone)]
+pub enum Qualifier {
+    Version(u64),
+    Alias(AliasName),
+    Latest,
+}
+
+impl Serialize for Qualifier {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        match self {
+            Qualifier::Version(x) => x.serialize(serializer),
+            Qualifier::Alias(a) => a.serialize(serializer),
+            Qualifier::Latest => "$LATEST".serialize(serializer),
+        }
+    }
+}
