+++ four/src/lambda/resource/mod.rs
@@ -6,7 +6,6 @@ pub mod layer_version;
 pub mod layer_version_permission;
 pub mod permission;
 pub mod url;
-pub mod utils;
 pub mod version;
 
 pub use alias::Alias;
