+++ four/src/lambda/resource/permission.rs
@@ -1,11 +1,8 @@
-use four_derive::ManagedResource;
-use nutype::nutype;
-use serde::Serialize;
-
 use crate::{
     core::{AccountDetail, AnyArn, LogicalId},
-    iam::{action::lambda::LambdaAction, ServicePrincipal},
-    lambda::{resource::url::AuthType, LooseFunctionName},
+    iam::action::lambda::LambdaAction,
+    lambda::{AuthType, EventSourceToken, LooseFunctionName, PermissionPrincipal, PrincipalOrgID},
+    ManagedResource,
 };
 
 #[derive(ManagedResource, Clone)]
@@ -16,42 +13,10 @@ pub struct Permission {
     event_source_token: Option<EventSourceToken>,
     function_name: LooseFunctionName,
     function_url_auth_type: Option<AuthType>,
-    principal: Option<Principal>,
+    principal: Option<PermissionPrincipal>,
 
     // TODO: rename (four_derive)
     principal_org_i_d: Option<PrincipalOrgID>,
     source_account: Option<AccountDetail>,
     source_arn: Option<AnyArn>,
 }
-
-#[nutype(
-    validate(not_empty, len_char_max = 256, regex = r#"^[a-zA-Z0-9._\-]+$"#),
-    derive(Debug, Clone, Serialize)
-)]
-pub struct EventSourceToken(String);
-
-#[derive(Debug, Clone)]
-enum Principal {
-    Service(ServicePrincipal),
-    Any,
-    Account(AccountDetail),
-}
-
-impl Serialize for Principal {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        match self {
-            Principal::Service(x) => x.serialize(serializer),
-            Principal::Any => "*".serialize(serializer),
-            Principal::Account(x) => x.serialize(serializer),
-        }
-    }
-}
-
-#[nutype(
-    validate(len_char_min = 12, len_char_max = 34, regex = r#"^o-[a-z0-9]{10,32}$"#),
-    derive(Debug, Clone, Serialize)
-)]
-pub struct PrincipalOrgID(String);
