+++ four/src/lambda/resource/url.rs
@@ -1,14 +1,10 @@
-use four_derive::ManagedResource;
-use nutype::nutype;
-use serde::Serialize;
-
 use crate::{
     core::{
         function::{HaveAtt, RefInner, Referenced},
-        service::Lambda,
-        Arn, LogicalId,
+        LogicalId,
     },
-    lambda::LooseFunctionName,
+    lambda::{AuthType, Cors, FunctionArn, InvokeMode, LooseFunctionName, Qualifier, UrlId},
+    ManagedResource,
 };
 
 #[derive(ManagedResource, Clone)]
@@ -22,72 +18,6 @@ pub struct Url {
     target_function_arn: LooseFunctionName,
 }
 
-#[derive(Debug, Clone, Serialize)]
-pub enum AuthType {
-    #[serde(rename(serialize = "AWS_IAM"))]
-    AwsIam,
-    #[serde(rename(serialize = "NONE"))]
-    None,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct Cors {
-    allow_credentials: Option<bool>,
-    allow_headers: Option<Vec<Header>>,
-    allow_methods: Option<Vec<Method>>,
-    allow_origins: Option<Vec<Origin>>,
-    expose_headers: Option<Vec<Header>>,
-    max_age: Option<MaxAge>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-#[serde(rename_all = "UPPERCASE")]
-pub enum Method {
-    Get,
-    Put,
-    Head,
-    Post,
-    Patch,
-    Delete,
-    #[serde(rename(serialize = "*"))]
-    Any,
-}
-
-#[derive(Debug, Clone, Serialize)]
-#[serde(untagged)]
-pub enum Origin {
-    #[serde(rename(serialize = "*"))]
-    Any,
-    Url(url::Url),
-}
-
-#[nutype(
-    validate(not_empty, len_char_max = 100),
-    derive(Debug, Clone, Serialize)
-)]
-pub struct Header(String);
-
-#[nutype(validate(less_or_equal = 86400), derive(Debug, Clone, Serialize))]
-pub struct MaxAge(usize);
-
-#[derive(Debug, Clone, Serialize)]
-#[serde(untagged)]
-pub enum InvokeMode {
-    #[serde(rename(serialize = "BUFFERED"))]
-    Buffered,
-    #[serde(rename(serialize = "RESPONSE_STREAM"))]
-    ResponseStream,
-}
-
-#[nutype(
-    validate(not_empty, len_char_max = 128, regex = r#"^[0-9][a-zA-Z0-9-_]*$"#),
-    derive(Debug, Clone, Serialize)
-)]
-pub struct Qualifier(String);
-
-#[derive(Debug, Clone, Serialize)]
-pub struct UrlId;
-
 impl Referenced for Url {
     type To = UrlId;
 
@@ -95,9 +25,6 @@ impl Referenced for Url {
         RefInner::Id(self.logical_id.clone())
     }
 }
-
-#[derive(Debug, Clone, Serialize)]
-pub struct FunctionArn(Arn<Lambda>);
 
 impl HaveAtt<FunctionArn> for Url {
     const KEY: &'static str = "FunctionArn";
